% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaca.R
\name{metaca}
\alias{metaca}
\title{Title  do meta-analysis by fitting random effect module}
\usage{
metaca(
  data,
  study = "Study",
  group = "Group",
  Control = "control",
  Case = "case",
  value = "Value",
  ci = 0.95
)
}
\arguments{
\item{data}{a data.frame which contain group information and values for statistic.}

\item{study}{a character string specifying the col.name of data, and the column contains data set information, default='Study'.}

\item{group}{a character string specifying the col.name of data, and the column contains group information, default='Group'}

\item{Control}{a character string specifying the control group.}

\item{Case}{a character string specifying the case group.}

\item{value}{a character string specifying the col.name of data which the column is the value for statistic,default='Value'.}

\item{ci}{confidence level of the interval, default=0.95.}
}
\value{
res a data.frame
}
\description{
Title  do meta-analysis by fitting random effect module
}
\examples{
metaca(data,study='Study',group='disease_stage',Control='healthy',Case='cancer',value='diversity')
}
