% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mataDA.R
\name{metaDA}
\alias{metaDA}
\title{Title}
\usage{
metaDA(data, md, n = 3)
}
\arguments{
\item{data}{OTU/gene/pathway abundance table}

\item{md}{metadata of each samples}

\item{n}{feature appears in more than n data sets}
}
\value{
DA a list of data.frame of results of each data sets and meta-analysis
}
\description{
Title
}
