% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_meta.R
\name{plot_meta}
\alias{plot_meta}
\title{Title Create a forest plot}
\usage{
plot_meta(res, size = 9)
}
\arguments{
\item{res}{a data.frame}

\item{size}{set fontsize}
}
\value{
a ggplot object
}
\description{
Title Create a forest plot
}
